feat: add ginseng benchmarks

+++ benchmark/benchmarks/ECSBenchmark.h
@@ -77,7 +77,7 @@ namespace ecs::benchmarks::base {
                 benchmark::ClobberMemory();
             }
             state.PauseTiming();
-            state.counters["entities"] = state.range(0);
+            state.counters["entities"] = static_cast<double>(state.range(0));
             this->uninitApplication(app);
             entities_factory.clear(app.getEntities());
         }
@@ -86,12 +86,12 @@ namespace ecs::benchmarks::base {
         void BM_DestroyEntities(benchmark::State& state) {
             std::vector<Entity> entities;
             Application app;
-            this->initApplicationWithEntities(app, state.range(0), entities);
+            this->initApplicationWithEntities(app, static_cast<size_t>(state.range(0)), entities);
             for (auto _ : state) {
                 state.PauseTiming();
                 entities.clear();
-                entities.reserve(state.range(0));
-                for (size_t i = 0; i < state.range(0); i++) {
+                entities.reserve(static_cast<size_t>(state.range(0)));
+                for (size_t i = 0; i < static_cast<size_t>(state.range(0)); i++) {
                     entities.push_back(entities_factory.createMinimal(app.getEntities()));
                 }
                 state.ResumeTiming();
@@ -103,7 +103,7 @@ namespace ecs::benchmarks::base {
                 benchmark::ClobberMemory();
             }
             state.PauseTiming();
-            state.counters["entities"] = entities.size();
+            state.counters["entities"] = static_cast<double>(entities.size());
             this->afterBenchmark(app);
             this->uninitApplication(app);
             entities.clear();
@@ -113,14 +113,14 @@ namespace ecs::benchmarks::base {
         void BM_UnpackOneComponent(benchmark::State& state) {
             std::vector<Entity> entities;
             Application app;
-            this->initApplicationWithEntities(app, state.range(0), entities);
+            this->initApplicationWithEntities(app, static_cast<size_t>(state.range(0)), entities);
             for (auto _ : state) {
                 for (auto& entity : entities) {
                     benchmark::DoNotOptimize(entities_factory.getComponentOne(app.getEntities(), entity));
                 }
             }
             state.PauseTiming();
-            state.counters["entities"] = entities.size();
+            state.counters["entities"] = static_cast<double>(entities.size());
             this->afterBenchmark(app);
             this->uninitApplication(app);
             entities.clear();
@@ -130,7 +130,7 @@ namespace ecs::benchmarks::base {
         void BM_UnpackTwoComponents(benchmark::State& state) {
             std::vector<Entity> entities;
             Application app;
-            this->initApplicationWithEntities(app, state.range(0), entities);
+            this->initApplicationWithEntities(app, static_cast<size_t>(state.range(0)), entities);
             for (auto _ : state) {
                 for (auto& entity : entities) {
                     benchmark::DoNotOptimize(entities_factory.getComponentOne(app.getEntities(), entity));
@@ -138,7 +138,7 @@ namespace ecs::benchmarks::base {
                 }
             }
             state.PauseTiming();
-            state.counters["entities"] = entities.size();
+            state.counters["entities"] = static_cast<double>(entities.size());
             this->afterBenchmark(app);
             this->uninitApplication(app);
             entities.clear();
@@ -149,7 +149,7 @@ namespace ecs::benchmarks::base {
             std::vector<Entity> entities;
             std::vector<Entity> entities_minimal;
             Application app;
-            this->initApplicationWithEntitiesAndMixedComponents(app, state.range(0), entities, entities_minimal);
+            this->initApplicationWithEntitiesAndMixedComponents(app, static_cast<size_t>(state.range(0)), entities, entities_minimal);
             for (auto _ : state) {
                 for (auto& entity : entities) {
                     benchmark::DoNotOptimize(entities_factory.getComponentOne(app.getEntities(), entity));
@@ -157,8 +157,8 @@ namespace ecs::benchmarks::base {
                 }
             }
             state.PauseTiming();
-            state.counters["entities"] = entities.size();
-            state.counters["entities_minimal"] = entities_minimal.size();
+            state.counters["entities"] = static_cast<double>(entities.size());
+            state.counters["entities_minimal"] = static_cast<double>(entities_minimal.size());
             this->afterBenchmark(app);
             this->uninitApplication(app);
             entities.clear();
@@ -170,7 +170,7 @@ namespace ecs::benchmarks::base {
             std::vector<Entity> entities;
             std::vector<Entity> entities_minimal;
             Application app (m_options.add_more_complex_system);
-            this->initApplicationWithEntitiesAndMixedComponents(app, state.range(0), entities, entities_minimal);
+            this->initApplicationWithEntitiesAndMixedComponents(app, static_cast<size_t>(state.range(0)), entities, entities_minimal);
             for (auto _ : state) {
                 for (auto& entity : entities) {
                     benchmark::DoNotOptimize(entities_factory.getComponentOne(app.getEntities(), entity));
@@ -179,8 +179,8 @@ namespace ecs::benchmarks::base {
                 }
             }
             state.PauseTiming();
-            state.counters["entities"] = entities.size();
-            state.counters["entities_minimal"] = entities_minimal.size();
+            state.counters["entities"] = static_cast<double>(entities.size());
+            state.counters["entities_minimal"] = static_cast<double>(entities_minimal.size());
             this->afterBenchmark(app);
             this->uninitApplication(app);
             entities.clear();
@@ -192,13 +192,13 @@ namespace ecs::benchmarks::base {
             std::vector<Entity> entities;
             std::vector<Entity> entities_minimal;
             Application app (m_options.add_more_complex_system);
-            this->initApplicationWithEntitiesAndMixedComponents(app, state.range(0), entities, entities_minimal);
+            this->initApplicationWithEntitiesAndMixedComponents(app, static_cast<size_t>(state.range(0)), entities, entities_minimal);
             for (auto _ : state) {
                 app.update(this->fakeTimeDelta);
             }
             state.PauseTiming();
-            state.counters["entities"] = entities.size();
-            state.counters["entities_minimal"] = entities_minimal.size();
+            state.counters["entities"] = static_cast<double>(entities.size());
+            state.counters["entities_minimal"] = static_cast<double>(entities_minimal.size());
             this->afterBenchmark(app);
             this->uninitApplication(app);
             entities.clear();
